+++ Tapjoy/TapjoyAdapterConfiguration.m
@@ -41,7 +41,7 @@ typedef NS_ENUM(NSInteger, TapjoyAdapterErrorCode) {
 #pragma mark - MPAdapterConfiguration
 
 - (NSString *)adapterVersion {
-    return @"12.2.1.0";
+    return @"12.3.1.0";
 }
 
 - (NSString *)biddingToken {