+++ Tapjoy/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+  * 12.3.1.0
+    * This version of adapters has been certified with Tapjoy 12.3.1.
+    * Fix misleading logging for ad load success and ad show failure in `TapjoyRewardedVideoCustomEvent`.
+    * Add `didClick` callback support for interstitial and rewarded video ad clicks.
+
   * 12.2.1.0
     * This version of the adapters has been certified with Tapjoy 12.2.1.
     * Pass MoPub's log level to Tapjoy. To adjust Tapjoy's log level via MoPub's log settings, reference the steps on [this page](https://developers.mopub.com/publishers/ios/test/#enable-logging).